spi: spi-xspi-if: Fix compiler warning of data type conversion
Add casting before XSPI_CDTBUF_TRTYPE to fix compiler warning of
data type conversion as below

 warning: conversion from ‘long unsigned int’ to ‘unsigned int’ changes
 value from ‘18446744073709518847’ to ‘4294934527’ [-Woverflow]

+++ drivers/memory/renesas-xspi-if.c
@@ -488,7 +488,7 @@ int xspi_manual_xfer(struct rpcif *xspi)
 			u32 nbytes, data[2], *p = data;
 
 			regmap_update_bits(xspi->regmap, XSPI_CDTBUF0,
-					XSPI_CDTBUF_TRTYPE, ~XSPI_CDTBUF_TRTYPE);
+					XSPI_CDTBUF_TRTYPE, ~(u32)XSPI_CDTBUF_TRTYPE);
 
 			/* nbytes may only be 1, 2, 4, or 8 */
 			nbytes = bytes_left >= max ? max : (1 << ilog2(bytes_left));
